Fix EOF errors when the mongo database gets disconnected.
Instead of calling Clone on existing sessions, we now call Copy, which
will acquire a new session from the pool and reconnect if necessary.
This prevents disconnected sockets from being reused if the mongo server
gets disconnected and also handles a new master server being elected
when part of a replicaset.

See:
https://github.com/go-mgo/mgo/issues/49#issuecomment-65122720
https://github.com/go-mgo/mgo/issues/47#issuecomment-63368722