Fix all numbers inserted or updated becoming strings.
The parsing of the JSON body was being handled by go-restful, which uses
the JSON decoder.UseNumber() method. This parses all numbers as strings
instead of integers or floats, which makes it impossible to insert or
update numeric values with mora. By handling the JSON parsing locally,
we can skip UseNumber, so integers and floats remain as those native
types.

See:
https://github.com/emicklei/go-restful/blob/v1.1.3/request.go#L112-L114