Use Mongo Extended JSON spec for outputing and querying BSON types.
This standardizes how complex BSON data types (like ObjectId, date
fields, binary data, etc) get translated to JSON. It uses MongoDB's
extended JSON specification:
http://docs.mongodb.org/manual/reference/mongodb-extended-json/ The
heavy lifting is done by the mejson go library:
https://github.com/compose/mejson

By treating the input queries as extended JSON, this allows for several
types of queries that weren't previously possible. For example, it's now
possible to query a date field like:

?query={"updated_at":{"$gt":{"$date":0}}}

Other special BSON types should be queryable in a similar fashion.

However, the output is backwards incompatible, since these various BSON
field types now have a different output style. For example, what might
have previously been returned as:

{
  "_id": "5505507a1c57d0372be0076c",
  "updated_at": "2015-03-15T09:27:22.889Z",
}

Is now returned as:

{
  "_id": {
    "$oid": "5505507a1c57d0372be0076c"
  },
  "updated_at": {
    "$date": 1426411642889
  }
}

String and numeric fields are not affected by this and will still be
returned the same.

The change in package paths for the mgo library as part of this commit
are for compatibility with the mejson library (which uses the gopkg.in
url instead of the github.com address).